+++ CHANGELOG.md
@@ -26,6 +26,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Change default limit of `read-outbox` command to 20 activities.
 - Change maximum length of local usernames to 30.
 - Update CAIP-10 account address regexp.
+- Present first object link in a post as a Misskey quote.
 
 ### Removed